requirements.txt
# ── Core ML ──────────────────────────────────────────────────────────────────
torch>=2.0.0
numpy>=1.24.0

# ── Scientific / metrics ────────────────────────────────────────────────────
scipy>=1.10.0
scikit-learn>=1.2.0

# ── Visualisation ───────────────────────────────────────────────────────────
matplotlib>=3.7.0

# ── Config / logging ────────────────────────────────────────────────────────
pyyaml>=6.0
tqdm>=4.65.0

# ── Landmark extraction (preprocessing only) ────────────────────────────────
mediapipe>=0.10.0
opencv-contrib-python>=4.7.0

# ── Tested with ─────────────────────────────────────────────────────────────
# torch==2.10.0  numpy==2.4.1  scipy==1.17.0  scikit-learn==1.8.0
# matplotlib==3.10.8  mediapipe==0.10.32  opencv-contrib-python==4.13.0.92
# PyYAML==6.0.3  tqdm==4.67.1  Python 3.13.11
